seminars/seminar4-event-driven/queues/worker/requirements.txt
pika==1.3.2
minio==7.2.0
pillow==10.1.0
requests==2.31.0
